requirements.txt
torchdiffeq
einops
numpy
pandas
scikit-learn
matplotlib
seaborn
wandb
tqdm
umap-learn
ipykernel
openpyxl
